(bug) ResourceSummary collection
When agents are deployed in managed cluster, collecting ResourceSummary
from the management cluster always caused an error to be logged (configMap
projectsveltos/drift-detection-version missing).

That is because ResourceSummary CRD is present in the management cluster
irrespective of whether there is a profile in drift detection mode is
matching the management cluster.

To fix this, irrespective of the mode, always collect clusters with
drift detection deployed and only collect resourceSummary instances from
those clusters.

This PR also fixes a crash when Sveltos is referencing a Flux Source and
the Flux source is deleted.