(bug) clean ResourceSummary instances when cluster is deleted
When in agentless, when a cluster is deleted, on top of removing
the drift detection deployment, also remove all resourceSummary
instances from that cluster.

This PR also optmizes collection of ResourceSummary instances
when in agentless mode. In this mode, all ResourceSummary instances
of all clusters are present in the management cluster.