(Bug Fix) Conflict Handover
When Profile A manages a set of Helm charts and Profile B
attempts to deploy one of the same charts, Sveltos correctly
identifies and reports a conflict.

However, if the chart is later removed from Profile A, Profile B
should be able to take over its management. A bug was preventing
this handover from occurring. This PR resolves that issue.