(bug) deployment error
When Sveltos deploys Kubernetes resources using _PolicyRefs_
or _KustomizationRefs_, it should only remove stale resources
**after** the new resources have been successfully deployed.

If deployment fails for any reason, stale resources should not
be removed.

For example, if Sveltos previously deployed resources A, B, and C,
and a new update (with invalid data) attempts to update resource
B but fails, Sveltos currently treats B and C as stale. This is
because it successfully redeploys A, fails on B, and assumes only
A is needed—leading to incorrect cleanup.