(bug) ResourceSummary use annotations for clusterSummary info
Deprecates `ClusterSummaryNamespaceLabel` and
`ClusterSummaryNameLabel`, replacing them with
`ClusterSummaryNamespaceAnnotation` and `ClusterSummaryNameAnnotation`.

This change addresses an issue where ResourceSummary creation failed
because ClusterSummary name exceeded the 63-character limit and so could not be
stored as a label value.

By switching to annotations, which do not have this restriction, we gain the
flexibility needed to handle longer names without errors.