(bug) Report failures during clusterSummary reconciliation
This PR introduces a new `failureMessage` field to the
ClusterSummary's top-level status.
This field will be used to report any errors encountered
during the ClusterSummary's reconciliation process itself,
distinct from errors that occur during the deployment
of individual features (which are reported in `featureSummaries`).

This enhancement provides greater visibility into reconciliation
failures that are not directly tied to feature deployment logic.