(bug) Ignore ErrReleaseNotFound
When Sveltos is installing a release, if an error happens before
retrying again, Sveltos might try to do an uninstall in order to
remove any stale helm Secret. This behavior was introduced because
while fixing this [bug](https://github.com/projectsveltos/addon-controller/issues/1204)

If uninstall fails with "release not found" ignore the error and proceed
with install.