(bug) ConfigurationBundles order in a ConfigurationGroup
The order of ConfigurationBundles is critical. This PR updates
the libsveltos dependency to ensure a required fix is included.
The fix addresses an issue where the order of ConfigurationBundles
within a ConfigurationGroup was not preserved.

Previously, when ConfigurationBundles were staged for a ConfigurationGroup,
Sveltos used a libsveltosset.Set to track them. However, since Set is backed
by a map, the order was unintentionally lost. The updated libsveltos replaces
the Set with a slice, preserving the original order.