(bug) make sure proper clean-up happens for helm charts in pull mode
In pull mode, the Addon Controller is responsible for preparing
the configurations (via ConfigurationGroup/ConfigurationBundles)
that the Sveltos Applier deploys. Once deployment is complete,
the Sveltos Applier updates the ConfigurationGroup status.

However, a potential issue can arise with Helm charts if a
ClusterSummary is deleted while the Sveltos Applier is still processing
a configuration.

For Helm, the undeployment process first involves creating a
ConfigurationGroup with Action: Deploy. This step is crucial for ensuring
that delete hooks are properly deployed and executed before the main
undeployment occurs.

Previously, if the Sveltos Applier was actively processing a configuration
and the ClusterSummary was then deleted, the Addon Controller could misinterpret
the applier's status update. It might have confused an update for a previous
configuration with an update for the delete hooks, potentially causing the delete hooks to be skipped.

This pull request addresses and resolves that specific issue.