Deployment to SveltosCluster in pull mode
SveltosCluster instances operating in pull mode necessitate
a valid license.
However, Sveltos offers a complimentary tier, deploying resources
to the initial two clusters configured for pull mode without a license.
Beyond this free limit, further deployments to pull-mode clusters will
cease, and an error will be logged in the Status of the respective
ClusterSummary instance.

It's important to note that this licensing requirement applies exclusively
to SveltosClusters in pull mode; all other functionalities will continue to
operate as usual.