(chore) standardizes helmChart chartName
This update standardizes how chartName is processed based on the Helm
registry type:

chartName must be in the format repositoryName/chartName.

If provided in this format, it is used as-is.

If chartName is provided without repositoryName prefixed, this change will
automatically prepend repositoryName/ to form the correct format.

chartName must not contain repositoryName.

If provided without repositoryName prefixed, it is used as-is.

If chartName is provided in the format repositoryName/chartName, this change
will automatically remove the repositoryName/ prefix, using chartName only.