(bug) fix delete stuck in pullmode
When in pull mode, if the profile is deleted (or stops matching a cluster)
while it is being deployed, Sveltos remains stuck with addon-controller
wanting to delete resources but the corresponding ConfigurationGroup with
action set to deploy.

This PR fixes that by deleting the ConfigurationGroup when in that state.

Fixes [51](https://github.com/projectsveltos/sveltos-applier/issues/51)