(bug) Drift-detection upgrade
When running with agents in the managed clusters, only upgrade drift-detection-manager
when ResourceSummary CRD is found in the managed cluster.
ResourceSummary CRD is deployed only when a profile is with SyncMode: ContinuousWithDriftDetection

Presence of the ResourceSummary CRD is an indication drift-detection manager needs to
be upgrade.

This bug was introduced in v1.3.0